Move delegates to separated domain (#124)
