+++ src/datasources/transaction-api/transaction-api.service.ts
@@ -9,7 +9,7 @@ import { Page } from '../../domain/entities/page.entity';
 import { MasterCopy } from '../../domain/chains/entities/master-copies.entity';
 import { Safe } from '../../domain/safe/entities/safe.entity';
 import { Contract } from '../../domain/contracts/entities/contract.entity';
-import { Delegate } from '../../domain/safe/entities/delegate.entity';
+import { Delegate } from '../../domain/delegate/entities/delegate.entity';
 
 function balanceCacheKey(chainId: string, safeAddress: string): string {
   return `${chainId}_${safeAddress}_balances`;
