+++ src/domain/delegate/entities/delegate.entity.ts
